ADD tests for balanced replay
