+++ tests/training/test_strategies.py
